Remove comment regarding LOCALIZATION2D
LOCALIZATION2D is not included in Slepian_Alpha, nor does it appear to be included in any of the other Slepian repositories.  I think it probably makes sense to do one of the following:
1. Remove the comment, as I have in this PR, or
2. Add a link to the script to the comment (http://geoweb.princeton.edu/people/simons/DOTM/localization2D.m), or
3. Add the script to Slepian_Alpha or one of the other Slepian repos.
I do not know which course of action is the most appropriate, so feel free to nix this PR if deleting the comment doesn't make sense in this case.